Split the models.py into rde/common_models.py, rde/rde_1_0_0.py
