+++ container_service_extension/client/cse_client/api_35/ovdc_api.py
@@ -6,10 +6,10 @@ from dataclasses import asdict
 
 import pyvcloud.vcd.client as vcd_client
 
+import container_service_extension.rde.models_.common_models
 from container_service_extension.client.cse_client.cse_client import CseClient
 from container_service_extension.client.response_processor import process_response  # noqa: E501
 import container_service_extension.common.constants.shared_constants as shared_constants  # noqa: E501
-import container_service_extension.rde.models as def_models
 
 
 class OvdcApi(CseClient):
@@ -40,9 +40,9 @@ class OvdcApi(CseClient):
             uri,
             self._client._session,
             accept_type='application/json')
-        return def_models.Ovdc(**process_response(response))
+        return container_service_extension.rde.models_.common_models.Ovdc(**process_response(response))
 
-    def update_ovdc(self, ovdc_id, ovdc_obj: def_models.Ovdc):
+    def update_ovdc(self, ovdc_id, ovdc_obj: container_service_extension.rde.models_.common_models.Ovdc):
         uri = f"{self._ovdc_uri}/{ovdc_id}"
         resp = self._client._do_request_prim(
             shared_constants.RequestMethod.PUT,
