+++ container_service_extension/client/de_cluster_native.py
@@ -9,6 +9,8 @@ import pyvcloud.vcd.exceptions as vcd_exceptions
 import yaml
 
 import container_service_extension.client.constants as cli_constants
+import container_service_extension.rde.models_.common_models
+import container_service_extension.rde.models_.rde_1_0_0
 from container_service_extension.client.cse_client.api_35.native_cluster_api import NativeClusterApi  # noqa: E501
 import container_service_extension.client.utils as client_utils
 import container_service_extension.common.constants.shared_constants as shared_constants  # noqa: E501
@@ -40,7 +42,7 @@ class DEClusterNative:
         self._native_cluster_api = NativeClusterApi(client)
         self._client = client
 
-    def create_cluster(self, cluster_entity: def_models.NativeEntity):
+    def create_cluster(self, cluster_entity: container_service_extension.rde.models_.rde_1_0_0.NativeEntity):
         """Create a new Kubernetes cluster.
 
         :param models.NativeEntity cluster_entity: native cluster entity
@@ -49,7 +51,7 @@ class DEClusterNative:
         msg = "Operation not supported; Under implementation"
         raise vcd_exceptions.OperationNotSupportedException(msg)
 
-    def resize_cluster(self, cluster_entity: def_models.NativeEntity):
+    def resize_cluster(self, cluster_entity: container_service_extension.rde.models_.rde_1_0_0.NativeEntity):
         """Resize the existing Kubernetes cluster.
 
         :param models.NativeEntity cluster_entity: native cluster entity
@@ -234,7 +236,7 @@ class DEClusterNative:
         :rtype: str
         """
         # TODO: check if we really need to decode-encode-decode-encode
-        cluster_upgrade_definition = def_models.DefEntity(**cluster_def_entity)
+        cluster_upgrade_definition = container_service_extension.rde.models_.common_models.DefEntity(**cluster_def_entity)
         cluster_def_entity = \
             self._native_cluster_api.upgrade_cluster_by_cluster_id(
                 cluster_id, cluster_upgrade_definition)
@@ -248,7 +250,7 @@ class DEClusterNative:
         :rtype: dict
         """
         entity_svc = def_entity_svc.DefEntityService(self._cloudapi_client)
-        cluster_spec = def_models.NativeEntity(**cluster_config)
+        cluster_spec = container_service_extension.rde.models_.rde_1_0_0.NativeEntity(**cluster_config)
         cluster_name = cluster_spec.metadata.cluster_name
         if cluster_id:
             # If cluster id doesn't exist, an exception will be raised
@@ -283,7 +285,7 @@ class DEClusterNative:
         # Parse user id info
         update_acl_entries = []
         for username, user_id in name_to_id.items():
-            acl_entry = def_models.ClusterAclEntry(
+            acl_entry = container_service_extension.rde.models_.common_models.ClusterAclEntry(
                 memberId=user_id,
                 username=username,
                 accessLevelId=access_level_id)
