+++ container_service_extension/server/request_handlers/v35/def_cluster_handler.py
@@ -3,6 +3,8 @@
 # SPDX-License-Identifier: BSD-2-Clause
 from dataclasses import asdict
 
+import container_service_extension.rde.models_.common_models
+import container_service_extension.rde.models_.rde_1_0_0
 from container_service_extension.common.constants.server_constants import CseOperation as CseServerOperationInfo  # noqa: E501
 from container_service_extension.common.constants.server_constants import FlattenedClusterSpecKey  # noqa: E501
 from container_service_extension.common.constants.server_constants import ThreadLocalData  # noqa: E501
@@ -16,7 +18,6 @@ import container_service_extension.common.utils.server_utils as server_utils
 import container_service_extension.lib.telemetry.constants as telemetry_constants  # noqa: E501
 import container_service_extension.lib.telemetry.telemetry_handler as telemetry_handler  # noqa: E501
 import container_service_extension.rde.cluster_service as cluster_svc
-import container_service_extension.rde.models as def_models
 import container_service_extension.security.context.operation_context as ctx
 import container_service_extension.server.request_handlers.request_utils as request_utils  # noqa: E501
 
@@ -32,7 +33,7 @@ def cluster_create(data: dict, op_ctx: ctx.OperationContext):
     :rtype: container_service_extension.def_.models.DefEntity
     """
     svc = cluster_svc.ClusterService(op_ctx)
-    cluster_entity_spec = def_models.NativeEntity(**data[RequestKey.INPUT_SPEC])  # noqa: E501
+    cluster_entity_spec = container_service_extension.rde.models_.rde_1_0_0.NativeEntity(**data[RequestKey.INPUT_SPEC])  # noqa: E501
     return asdict(svc.create_cluster(cluster_entity_spec))
 
 
@@ -48,7 +49,7 @@ def cluster_resize(data: dict, op_ctx: ctx.OperationContext):
     """
     svc = cluster_svc.ClusterService(op_ctx)
     cluster_id = data[RequestKey.CLUSTER_ID]
-    cluster_entity_spec = def_models.NativeEntity(**data[RequestKey.INPUT_SPEC])  # noqa: E501
+    cluster_entity_spec = container_service_extension.rde.models_.rde_1_0_0.NativeEntity(**data[RequestKey.INPUT_SPEC])  # noqa: E501
     curr_entity = svc.entity_svc.get_entity(cluster_id)
     request_utils.validate_request_payload(
         asdict(cluster_entity_spec.spec), asdict(curr_entity.entity.spec),
@@ -126,7 +127,7 @@ def cluster_upgrade(data, op_ctx: ctx.OperationContext):
     :return: Dict
     """
     svc = cluster_svc.ClusterService(op_ctx)
-    cluster_entity_spec = def_models.NativeEntity(**data[RequestKey.INPUT_SPEC])  # noqa: E501
+    cluster_entity_spec = container_service_extension.rde.models_.rde_1_0_0.NativeEntity(**data[RequestKey.INPUT_SPEC])  # noqa: E501
     cluster_id = data[RequestKey.CLUSTER_ID]
     curr_entity = svc.entity_svc.get_entity(cluster_id)
     request_utils.validate_request_payload(
