+++ container_service_extension/server/service.py
@@ -23,6 +23,7 @@ import container_service_extension.common.constants.shared_constants as shared_c
 import container_service_extension.common.utils.core_utils as utils
 import container_service_extension.common.utils.pyvcloud_utils as vcd_utils
 import container_service_extension.common.utils.server_utils as server_utils
+import container_service_extension.rde.models_.common_models
 from container_service_extension.common.utils.vsphere_utils import populate_vsphere_list  # noqa: E501
 import container_service_extension.exception.exceptions as cse_exception
 import container_service_extension.installer.configure_cse as configure_cse
@@ -39,7 +40,6 @@ from container_service_extension.mqi.consumer.consumer import MessageConsumer
 from container_service_extension.mqi.mqtt_extension_manager import \
     MQTTExtensionManager
 import container_service_extension.rde.constants as def_constants
-import container_service_extension.rde.models as def_models
 import container_service_extension.rde.schema_service as def_schema_svc
 import container_service_extension.rde.utils as def_utils
 from container_service_extension.rde.utils import raise_error_if_def_not_supported  # noqa: E501
@@ -184,8 +184,8 @@ class Service(object, metaclass=Singleton):
         self.skip_config_decryption = skip_config_decryption
         self.pks_cache = None
         self._state = ServerState.STOPPED
-        self._kubernetesInterface: def_models.DefInterface = None
-        self._nativeEntityType: def_models.DefEntityType = None
+        self._kubernetesInterface: container_service_extension.rde.models_.common_models.DefInterface = None
+        self._nativeEntityType: container_service_extension.rde.models_.common_models.DefEntityType = None
         self.consumer = None
         self.consumer_thread = None
         self._consumer_watchdog = None
@@ -226,11 +226,11 @@ class Service(object, metaclass=Singleton):
             del result['python']
         return result
 
-    def get_kubernetes_interface(self) -> def_models.DefInterface:
+    def get_kubernetes_interface(self) -> container_service_extension.rde.models_.common_models.DefInterface:
         """Get the built-in kubernetes interface from vCD."""
         return self._kubernetesInterface
 
-    def get_native_cluster_entity_type(self) -> def_models.DefEntityType:
+    def get_native_cluster_entity_type(self) -> container_service_extension.rde.models_.common_models.DefEntityType:
         return self._nativeEntityType
 
     def update_status(self, server_action: shared_constants.ServerAction):
